setup for strategy Less Spool Changes

+++ src/main/java/nl/saxion/models/prints/Print.java
@@ -33,7 +33,7 @@ public class Print {
         return name;
     }
 
-    public double getLength() {
+    public int getLength() {
         return length;
     }
